TIP-706: Move resetIndex into the akeneo TestCase Class
