+++ src/Pim/Bundle/CatalogBundle/tests/integration/PQB/Filter/IdentifierFilterIntegration.php
@@ -20,8 +20,6 @@ class IdentifierFilterIntegration extends AbstractProductQueryBuilderTestCase
         parent::setUp();
 
         if (1 === self::$count || $this->getConfiguration()->isDatabasePurgedForEachTest()) {
-            $this->resetIndex();
-
             $this->createProduct('foo', []);
             $this->createProduct('bar', []);
             $this->createProduct('baz', []);
