+++ src/Pim/Bundle/CatalogBundle/tests/integration/PQB/Filter/Media/LocalizableScopableFilterIntegration.php
@@ -20,8 +20,6 @@ class LocalizableScopableFilterIntegration extends AbstractProductQueryBuilderTe
         parent::setUp();
 
         if (1 === self::$count || $this->getConfiguration()->isDatabasePurgedForEachTest()) {
-            $this->resetIndex();
-
             $this->createProduct('product_one', [
                 'values' => [
                     'a_localizable_scopable_image' => [
