+++ src/Pim/Bundle/CatalogBundle/tests/integration/PQB/Filter/Options/LocalizableFilterIntegration.php
@@ -21,8 +21,6 @@ class LocalizableFilterIntegration extends AbstractProductQueryBuilderTestCase
         parent::setUp();
 
         if (1 === self::$count || $this->getConfiguration()->isDatabasePurgedForEachTest()) {
-            $this->resetIndex();
-
             $this->createAttribute([
                 'code'                => 'a_localizable_multi_select',
                 'type'                => AttributeTypes::OPTION_MULTI_SELECT,
