+++ src/Pim/Bundle/CatalogBundle/tests/integration/PQB/Filter/Options/ScopableFilterIntegration.php
@@ -21,8 +21,6 @@ class ScopableFilterIntegration extends AbstractProductQueryBuilderTestCase
         parent::setUp();
 
         if (1 === self::$count || $this->getConfiguration()->isDatabasePurgedForEachTest()) {
-            $this->resetIndex();
-
             $this->createAttribute([
                 'code'                => 'a_scopable_multi_select',
                 'type'                => AttributeTypes::OPTION_MULTI_SELECT,
