+++ src/Pim/Bundle/CatalogBundle/tests/integration/PQB/Filter/Price/LocalizableScopableFilterIntegration.php
@@ -21,8 +21,6 @@ class LocalizableScopableFilterIntegration extends AbstractProductQueryBuilderTe
         parent::setUp();
 
         if (1 === self::$count || $this->getConfiguration()->isDatabasePurgedForEachTest()) {
-            $this->resetIndex();
-
             $this->createAttribute([
                 'code'             => 'a_scopable_localizable_price',
                 'type'             => AttributeTypes::PRICE_COLLECTION,
