+++ src/Pim/Bundle/CatalogBundle/tests/integration/PQB/Filter/StatusFilterIntegration.php
@@ -20,8 +20,6 @@ class StatusFilterIntegration extends AbstractProductQueryBuilderTestCase
         parent::setUp();
 
         if (1 === self::$count || $this->getConfiguration()->isDatabasePurgedForEachTest()) {
-            $this->resetIndex();
-
             $this->createProduct('foo', ['enabled' => true]);
             $this->createProduct('bar', ['enabled' => false]);
         }
