+++ src/Pim/Bundle/CatalogBundle/tests/integration/PQB/Filter/Text/LocalizableScopableFilterIntegration.php
@@ -21,8 +21,6 @@ class LocalizableScopableFilterIntegration extends AbstractProductQueryBuilderTe
         parent::setUp();
 
         if (1 === self::$count || $this->getConfiguration()->isDatabasePurgedForEachTest()) {
-            $this->resetIndex();
-
             $this->createAttribute([
                 'code'                => 'a_localizable_scopable_text',
                 'type'                => AttributeTypes::TEXT,
