+++ src/Pim/Bundle/CatalogBundle/tests/integration/PQB/Filter/Text/TextFilterIntegration.php
@@ -20,8 +20,6 @@ class TextFilterIntegration extends AbstractProductQueryBuilderTestCase
         parent::setUp();
 
         if (1 === self::$count || $this->getConfiguration()->isDatabasePurgedForEachTest()) {
-            $this->resetIndex();
-
             $this->createProduct('cat', [
                 'values' => [
                     'a_text' => [['data' => 'cat', 'locale' => null, 'scope' => null]],
